Move group store back to internal package; refactor from using group store to passing GetFiltered() function
