+++ central/role/datastore/datastore_impl_test.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"testing"
 
-	"github.com/golang/mock/gomock"
-	groupMock "github.com/stackrox/rox/central/group/datastore/store/mocks"
 	"github.com/stackrox/rox/central/role"
 	"github.com/stackrox/rox/central/role/resources"
 	"github.com/stackrox/rox/central/role/store"
@@ -80,8 +78,6 @@ type roleDataStoreTestSuite struct {
 	boltDB    *bolt.DB
 	rocksie   *rocksdb.RocksDB
 
-	groupStore *groupMock.MockStore
-
 	postgresTest *pgtest.TestPostgres
 
 	existingRole                     *storage.Role
@@ -89,6 +85,9 @@ type roleDataStoreTestSuite struct {
 	existingScope                    *storage.SimpleAccessScope
 	existingDeclarativePermissionSet *storage.PermissionSet
 	existingDeclarativeScope         *storage.SimpleAccessScope
+
+	filteredFuncReturnValue []*storage.Group
+	filteredFuncReturnError error
 }
 
 func (s *roleDataStoreTestSuite) SetupTest() {
@@ -106,6 +105,10 @@ func (s *roleDataStoreTestSuite) SetupTest() {
 	s.hasWriteDeclarativeCtx = declarativeconfig.WithModifyDeclarativeResource(s.hasWriteCtx)
 
 	s.initDataStore()
+}
+
+func (s *roleDataStoreTestSuite) mockGroupGetFiltered(ctx context.Context, filter func(*storage.Group) bool) ([]*storage.Group, error) {
+	return s.filteredFuncReturnValue, s.filteredFuncReturnError
 }
 
 func (s *roleDataStoreTestSuite) initDataStore() {
@@ -132,8 +135,7 @@ func (s *roleDataStoreTestSuite) initDataStore() {
 		s.Require().NoError(err)
 	}
 
-	s.groupStore = groupMock.NewMockStore(gomock.NewController(s.T()))
-	s.dataStore = New(roleStorage, permissionSetStorage, accessScopeStorage, s.groupStore)
+	s.dataStore = New(roleStorage, permissionSetStorage, accessScopeStorage, s.mockGroupGetFiltered)
 
 	// Insert a permission set, access scope, and role into the test DB.
 	s.existingPermissionSet = getValidPermissionSet("permissionset.existing", "existing permissionset")
@@ -277,7 +279,7 @@ func (s *roleDataStoreTestSuite) TestRoleWriteOperations() {
 	badDeclarativeRole.Traits = &storage.Traits{
 		Origin: storage.Traits_DECLARATIVE,
 	}
-	s.groupStore.EXPECT().Walk(gomock.Any(), gomock.Any()).Return(nil).AnyTimes()
+	s.setupGetFilteredReturnValues([]*storage.Group{}, nil)
 
 	err := s.dataStore.AddPermissionSet(s.hasWriteCtx, secondExistingPermissionSet)
 	s.NoError(err, "failed to add second permission set needed for test")
@@ -380,6 +382,11 @@ func (s *roleDataStoreTestSuite) TestRoleWriteOperations() {
 
 	err = s.dataStore.UpsertRole(s.hasWriteDeclarativeCtx, badDeclarativeRole)
 	s.ErrorIs(err, errox.InvalidArgs, "invalid scope for Upsert*() yields an error(declarative resource)")
+}
+
+func (s *roleDataStoreTestSuite) setupGetFilteredReturnValues(groups []*storage.Group, err error) {
+	s.filteredFuncReturnValue = groups
+	s.filteredFuncReturnError = err
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -897,15 +904,15 @@ func (s *roleDataStoreTestSuite) TestForeignKeyConstraints() {
 	err = s.dataStore.RemoveAccessScope(s.hasWriteCtx, scope.GetId())
 	s.ErrorIs(err, errox.ReferencedByAnotherObject, "cannot delete an Access Scope referred to by a Role")
 
-	s.groupStore.EXPECT().Walk(gomock.Any(), gomock.Any()).Do(func(ctx context.Context, f func(group *storage.Group) error) {
-		f(&storage.Group{
+	s.setupGetFilteredReturnValues([]*storage.Group{
+		{
 			RoleName: role.GetName(),
-		})
-	}).Return(nil)
+		},
+	}, nil)
 	err = s.dataStore.RemoveRole(s.hasWriteCtx, role.GetName())
 	s.ErrorIs(err, errox.ReferencedByAnotherObject)
 
-	s.groupStore.EXPECT().Walk(gomock.Any(), gomock.Any()).Return(nil)
+	s.setupGetFilteredReturnValues([]*storage.Group{}, nil)
 	err = s.dataStore.RemoveRole(s.hasWriteCtx, role.GetName())
 	s.NoError(err)
