+++ central/role/datastore/datastore_test_constructors.go
@@ -3,8 +3,7 @@ package datastore
 import (
 	"testing"
 
-	"github.com/golang/mock/gomock"
-	groupMock "github.com/stackrox/rox/central/group/datastore/store/mocks"
+	groupFilter "github.com/stackrox/rox/central/group/datastore/filter"
 	permissionSetPostgresStore "github.com/stackrox/rox/central/role/store/permissionset/postgres"
 	permissionSetRocksDBStore "github.com/stackrox/rox/central/role/store/permissionset/rocksdb"
 	rolePostgresStore "github.com/stackrox/rox/central/role/store/role/postgres"
@@ -20,9 +19,8 @@ func GetTestPostgresDataStore(t *testing.T, pool *postgres.DB) (DataStore, error
 	permissionStore := permissionSetPostgresStore.New(pool)
 	roleStore := rolePostgresStore.New(pool)
 	scopeStore := accessScopePostgresStore.New(pool)
-	groupStore := groupMock.NewMockStore(gomock.NewController(t))
 
-	return New(roleStore, permissionStore, scopeStore, groupStore), nil
+	return New(roleStore, permissionStore, scopeStore, groupFilter.GetFiltered), nil
 }
 
 // GetTestRocksBleveDataStore provides a datastore connected to rocksdb and bleve for testing purposes.
@@ -39,7 +37,6 @@ func GetTestRocksBleveDataStore(t *testing.T, rocksengine *rocksdbBase.RocksDB) 
 	if err != nil {
 		return nil, err
 	}
-	groupStore := groupMock.NewMockStore(gomock.NewController(t))
 
-	return New(roleStore, permissionStore, scopeStore, groupStore), nil
+	return New(roleStore, permissionStore, scopeStore, groupFilter.GetFiltered), nil
 }
